ci: temporarily disable crate publish
At the moment, unfortunately there is a high probability that the GHA agent runs out of disk space
during the publish. I used to have this running on a custom agent that ran on AWS, but that randomly
stopped working without much apparent reason.

For now, I will just publish manually.