chore!: remove the `--testnet` argument
BREAKING CHANGE: the `ant`, `antctl` and `antnode` binaries will no longer support the `--testnet`
argument.

This argument will no longer be required because we can designate we are using a testnet by setting
`--network-id` to any value above 2, where 1 will be reserved for `mainnet` and 2 will be reserved
for `alphanet`.

A test case is added to ensure that a serialized node service data struct can still be deserialized
when the `disable_mainnet_contacts` field is missing.