feat: provide `--alpha` argument
The `ant`, `antnode` and `antctl` binaries will all now support an `--alpha` argument as part of
their peer configurations.

When the argument is used, a network ID of 2 will be applied, and the initial peers will be fetched
from a set of five static IP addresses that have been allocated for the alpha network. So if users
want to connect to this network, they need only use this switch rather than manually supplying
`--network-id`, `--network-contacts-url` and/or `--peer` arguments.