chore: formatting and test fix up
I still don't have Cursor setup like Neovim where it formatted on save.

This commit also fixes up a few bootstrapping tests after the code was changed. The code will return
the production network contacts unless it returns early, and that is done by putting a limit on the
number of contacts.