chore: explicitly set version strings when setting network id
It is possible that some of these strings would be read before the network ID is set and they would
then have their value assigned for the lifetime of the program, even after `set_network_id` is
called.

They all need to be explicitly updated when `set_network_id` is called.