chore: resolve clippy warnings for rust 1.87
Warnings for large enum variants and returned results were suppressed at crate level for a few
crates because I'm trying to get a hotfix released and I didn't have time to change the code to
resolve all the warnings.

A test also had to be modified because the results it was expecting were now different. Possibly
something to do with floating point precision.