feat: support network id from 0-255
There was a bit of a misunderstanding in the way the `NetworkId` was implemented. It was reserving
the values 0, 1, and 2, for local, main, and alpha networks, respectively, then assigning 3 for any
other 'custom' network.

We change it here such that 0, 1, and 2 are reserved as above, but custom networks are permitted to
be between 3 and 255.