ci: provide tests for covering all file types
For both small and medium files:
* public without archive
* public with archive
* public directory
* public directory without archive
* private without archive
* private with archive
* private directory
* private directory without archive

Where a small file is something between 1 to 50MB, and a medium file is between 50MB and 1GB.

Since these are bit cumbersome to have in shell, it's arguable they could have been implemented as
Rust tests for `autonomi`. However, things like the `--no-archive` flag is a feature of the CLI, not
the `autonomi` API.

I also chose to implement these as a separate job from the e2e test, and just leave that as
something that deals with smaller cases for each feature.