chore: only build specified binaries on release
When triggering the workflow the user will get a chance to disable specific binaries from being
built, in which case the current version of the binary will be used.

The reason is because it's possible there can be code changes for binaries whose versions we chose
not to bump. If there was no version bump, we don't want the binary to be built, either because
there are no changes or we don't want any new code to be released at this point.

The person who is composing the release should be aware of which binaries need to be built, rather
than having some automated mechanism to figure it out.