chore: remove various nightly workflows
This resolves issue #3209 reported by a community member.

In addition to removing that specific workflow, the other nightly workflows were also removed. None
of these were being actively used. If we ever want to run a set of nightly tests again, these can be
retrieved from the version control system if need be.