feat: `antnode` automatic upgrades for unix platforms
Automatic upgrades aim to make it as easy as possible for node operators to obtain new versions of
the `antnode` binary. We also provide a 'staggering' mechanism to attempt to avoid the turbulence
caused when everyone upgrades around the same time. If the majority of operators use this feature,
we should see better network health and stability, because newer versions of the node will propagate
quicker and safer.

Initially, we only support Linux and macOS. The Windows operating system does not allow a binary
associated with a running process to be overwritten, which is part of the upgrade process (see
below).

An upgrade is performed by a background thread that wakes up every 72 hours.

After the upgrade is performed, the restart of the binary is delayed for a period of time that is
determined by taking the hash of the peer ID and the size of the network into account, with a cap of
72 hours. This is to make the upgrades 'staggered', to avoid many thousands of node processes
restarting at the same time. To facilitate the calculation, the `get_kbuckets` API is slightly
modified to return the estimated network size.

The upgrade process works as follows:

* Get the latest release information from the `autonomi` repository.
* If the commit hash of the latest release matches the commit hash of the currently running binary,
  do nothing and the process ends. Otherwise the upgrade proceeds.
* Get the currently running platform. This information is used to locate the correct binary and
  sha256 hash.
* Check if another `antnode` process has already downloaded the new binary to a temporary location
  and its sha256 hash matches that of the binary from the fetched release information. This will
  prevent many `antnode` processes continually downloading the new binary. If the new binary has not
  already been downloaded, it will be downloaded and extracted to the temporary location.
* The hash of the downloaded binary will be compared to the hash of the binary in the new release
  information.
* The binary of the currently running process will be overwritten by the new binary. We will check
  to see if the hash of the running binary matches the hash of the upgrade binary; in this case,
  another process already upgraded the binary and we don't need to do anything.

If an upgrade occurred, the `antnode` process will be restarted after a delay of some period, as
mentioned above.

Given the more elaborate implementation of automatic upgrades, the workflow for self restarting was
deleted.