chore: upgrade `assert_cmd` reference
When a `cargo update` was being performed, it appeared to be pulling in a newer version of this
`assert_cmd` crate, and this newer version had deprecated a function we were using in some tests,
which in turn was causing Clippy to raise some warnings.

Here we update to the latest version of `assert_cmd` and remove the use of the deprecated function
in favour of a new macro, which resolves the Clippy warnings.