feat: improve process detection for antnode services
The upgrade process introduced a problem with the detection of processes using the path-based
approach we had. When the binary is overwritten, it seems the text " (deleted)" gets appended to the
path, and thus the `antctl status` command was marking services as stopped when they were still
running and scheduled to be upgraded.

This new approach first tries to use the stored PID, then falls back on the path-detection
mechanism, attempting to take the "(deleted)" suffix into account.