fix: do not truncate upgrade lock file
By examining the logs on a testnet run I determined all the `antnode` processes on the same
host downloaded the upgrade binary. This is something we expressly wanted to avoid.

The problem was the `File::create` function was truncating the lock file. It has now been changed to
avoid this behaviour, so other processes should wait to acquire the lock.