feat: provide cache for release info
In a similar fashion to the cache for the upgrade binary, we provide a cache for the release
information.

While performing a test with 5000 nodes I got rate limited by Github when thousands of processes
queried for the release information at approximately the same time. Now it should only be one query
per host.