chore: use sha256 hash in restart delay calculation
In the test run with 1000 nodes we observed a distribution of restart times that seemed a bit too
tightly clustered.

Using this alternative hashing mechanism resulted in a distribution that was more even.

The restart delay calculation is also moved outside the loop because it only needs to be calculated
once.