chore: use 'on success' restart policy
Using 'on success' rather than 'always' allows the service manager to restart the service when it
exits after an upgrade, but stop it when it exits on an error.

This should allow us to support automatic upgrades in `node-launchpad` without compromising the
behaviour we need to retain for stopping the node when UPnP is not working.