feat: support restarting symlinked binaries
The code performing restarts of `antnode` was using the canonical path of the executable, which in
the case of symlinks, is the path of the source of the link rather than the link itself.

To support a symlinked binary setup in automatic upgrades, we now modify the restart mechanism to
use the invoked path, which is the location of the symlink. For non-symlinked binaries the mechanism
should work as normal.

The way the new `antnode` process is launched was also changed slightly. We were using a
`retain_peer_id` flag, which was supplied from the RPC command. Since RPC will be phased out, this
flag is no longer relevant and was removed. The restart will now always retain the peer ID. If the
original process supplied the `--root-dir` and `--port` arguments, we'll use them again, but if it
didn't, they will be injected. Now they need to be injected at the beginning rather than appended at
the end, because the EVM options must be at the end.