feat: assign random port for node when adding service
The `antctl add` command has an optional `--node-port` argument. In its absence `antnode` will
assign the port randomly. Now this behaviour is changed to have the `add` command select a random
port and emit the `--port` argument when writing the service definition.

The reason for this is so that nodes will always restart with the same port after being upgraded
automatically. If it wasn't done this way, the service manager would restart the service using the
same arguments as before, which wouldn't have the `--port` argument in them, and thus the node would
then select a new random port.