fix: macos services stop on `anctl stop`
The new version of the underlying service manager crate was incorrectly specifying an 'on success'
restart policy which was preventing services from being stopped on macOS when the `anctl stop`
command was used. This has now been fixed in the underlying crate.

The restart policy uses the `KeepAlive` setting, which was required for making the services
compatible with the automatic upgrade feature. This also had an unintended side effect where it
caused services to start as soon as they were added with the `antctl add` command, which is
unintuitive and not the way other service managers behave.

The new version of the crate also fixes this issue.